requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.0
django-etc==1.4.0
django-filter==23.3
django-model-utils==4.3.1
djangorestframework==3.14.0
Faker==20.0.0
idna==3.4
Markdown==3.5.1
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
validators==0.22.0
